fix: Update Engine to pass dt parameter to player
The engine file wasn't passing the dt parameter to the player, this is necessary for the update method for both enemies and player and it is expected in the app file. The dt parameter ensure the game runs at the same speed for all computers and it would feel awkward if the enemies had it and the player didn't.